correção controle de armazenamento

+++ backend/src/database/migrations/20260422140000-company-storage-snapshots-and-watermark.ts
@@ -46,7 +46,8 @@ module.exports = {
       )
       .then(() =>
         queryInterface.addColumn("Companies", "storageAlertWatermark", {
-          type: DataTypes.TINYINT,
+          // SMALLINT: compatível com PostgreSQL e MySQL (PostgreSQL não tem TINYINT).
+          type: DataTypes.SMALLINT,
           allowNull: false,
           defaultValue: 0
         })
